componentized remaining pages, disabled router direction animation
